i just added the welcome Dialog message in the app
