+++ app/src/main/java/com/project/ksih_android/ui/auth/LoginFragment.java
@@ -1,7 +1,10 @@
 package com.project.ksih_android.ui.auth;
 
+import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.preference.PreferenceManager;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -21,11 +24,16 @@ import com.victor.loading.rotate.RotateLoading;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
 import androidx.databinding.DataBindingUtil;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
 import androidx.navigation.Navigation;
+
+import java.util.Timer;
+import java.util.TimerTask;
+
 import timber.log.Timber;
 
 /**
@@ -37,6 +45,7 @@ public class LoginFragment extends Fragment {
     private LoginViewModel mViewModel;
     private FirebaseAuth mAuth;
     private FragmentLoginBinding mLoginBinding;
+
 
     @Nullable
     @Override
@@ -79,32 +88,30 @@ public class LoginFragment extends Fragment {
 
     private void signInUser(final String email, String password) {
         mAuth.signInWithEmailAndPassword(email, password)
-                .addOnCompleteListener(requireActivity(), new OnCompleteListener<AuthResult>() {
-            @Override
-            public void onComplete(@NonNull Task<AuthResult> task) {
-                if (task.isSuccessful()) {
-                    Timber.d("Log in successful");
-                    if (!isUserVerified(mAuth.getCurrentUser())) {
-                        Toast.makeText(getActivity(), "Your Email has not been verified", Toast.LENGTH_SHORT).show();
-                        stopProgressBar(mLoginBinding.progressBar);
-                        showButton(mLoginBinding.buttonSignIn);
-                        logout();
-                        Timber.d("UserNotVerified");
+                .addOnCompleteListener(requireActivity(), task -> {
+                    if (task.isSuccessful()) {
+                        Timber.d("Log in successful");
+                        if (!isUserVerified(mAuth.getCurrentUser())) {
+                            Toast.makeText(getActivity(), "Your Email has not been verified", Toast.LENGTH_SHORT).show();
+                            stopProgressBar(mLoginBinding.progressBar);
+                            showButton(mLoginBinding.buttonSignIn);
+                            logout();
+                            Timber.d("UserNotVerified");
+                        } else {
+                            Timber.d("UserIsVerified");
+                            stopProgressBar(mLoginBinding.progressBar);
+                            showButton(mLoginBinding.buttonSignIn);
+                            Toast.makeText(getActivity(), "Sign in successful", Toast.LENGTH_SHORT).show();
+                            navigateToHomeActivity();
+                            isFirstTimeLogin();
+                        }
                     } else {
-                        Timber.d("UserIsVerified");
                         stopProgressBar(mLoginBinding.progressBar);
                         showButton(mLoginBinding.buttonSignIn);
-                        Toast.makeText(getActivity(), "Sing in successful", Toast.LENGTH_SHORT).show();
-                        navigateToHomeActivity();
+                        Toast.makeText(getActivity(), task.getException().getLocalizedMessage(), Toast.LENGTH_LONG).show();
+                        Timber.d("SignInError: " + task.getException().getMessage());
                     }
-                } else {
-                    stopProgressBar(mLoginBinding.progressBar);
-                    showButton(mLoginBinding.buttonSignIn);
-                    Toast.makeText(getActivity(), task.getException().getLocalizedMessage(), Toast.LENGTH_LONG).show();
-                    Timber.d("SignInError: " + task.getException().getMessage());
-                }
-            }
-        });
+                });
     }
 
     private boolean isUserVerified(FirebaseUser user) {
@@ -118,7 +125,14 @@ public class LoginFragment extends Fragment {
     }
 
     private void navigateToHomeActivity() {
-        startActivity(new Intent(requireActivity(), HomeActivity.class));
+        new Timer().schedule(new TimerTask() {
+            @Override
+            public void run() {
+                startActivity(new Intent(requireActivity(), HomeActivity.class));
+            }
+        }, 8000);
+
+
     }
 
     private void navigateToRegisterFragment(TextView textView) {
@@ -153,5 +167,37 @@ public class LoginFragment extends Fragment {
 
     private void hideButton(MaterialButton button) {
         button.setVisibility(View.INVISIBLE);
+    }
+
+    public void isFirstTimeLogin() {
+
+        final SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(getActivity());
+        boolean firstTimeLogin = preferences.getBoolean("First Login", true);
+
+        if (firstTimeLogin) {
+
+            AlertDialog.Builder messageBuilder = new AlertDialog.Builder(getActivity());
+            messageBuilder.setMessage(getString(R.string.first_time_welcoming_message));
+
+            messageBuilder.setPositiveButton("yes", new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(DialogInterface dialog, int which) {
+                    SharedPreferences.Editor messageEditor = preferences.edit();
+                    messageEditor.putBoolean("First Login", false);
+                    messageEditor.commit();
+                    dialog.dismiss();
+
+                }
+            });
+
+            messageBuilder.setIcon(R.drawable.ksih_background);
+            messageBuilder.setTitle(" ");
+            AlertDialog alertDialogKsih = messageBuilder.create();
+            alertDialogKsih.show();
+
+
+        }
+
+
     }
 }
