+++ app/src/main/java/com/project/ksih_android/ui/ksihrules/KSIHRules.java
@@ -3,6 +3,7 @@ package com.project.ksih_android.ui.ksihrules;
 
 import android.os.Bundle;
 
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.databinding.DataBindingUtil;
 import androidx.fragment.app.Fragment;
@@ -14,6 +15,11 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
 import com.project.ksih_android.R;
 import com.project.ksih_android.databinding.FragmentKsihrulesBinding;
 import com.project.ksih_android.ui.member.MemberViewModel;
@@ -26,6 +32,9 @@ public class KSIHRules extends Fragment {
 
 
     private KSIHRulesViewModel mKSIHRulesViewModel;
+    private String ksihRules;
+    private FirebaseDatabase mFirebaseDatabaseKsihRules;
+    private DatabaseReference mDatabaseReferenceKsihRules;
 
 
     public KSIHRules() {
@@ -38,15 +47,22 @@ public class KSIHRules extends Fragment {
         mKSIHRulesViewModel = ViewModelProviders.of(this).get(KSIHRulesViewModel.class);
         // Inflate the layout for this fragment
         mFragmentKsihrulesBinding = DataBindingUtil.inflate(inflater, R.layout.fragment_ksihrules, container, false);
-        mFragmentKsihrulesBinding.getLifecycleOwner();
-
-        mKSIHRulesViewModel.getText().observe(this, new Observer<String>() {
+        mDatabaseReferenceKsihRules = mFirebaseDatabaseKsihRules.getReference("ksih_rules");
+        mDatabaseReferenceKsihRules.addValueEventListener(new ValueEventListener() {
             @Override
-            public void onChanged(@Nullable String s) {
-                mFragmentKsihrulesBinding.textviewKsihRules.setText("Ksih rules");
-                //             mFragmentKsihrulesBinding.textviewKsihRules.setText(s);
+            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+
+
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError databaseError) {
+
             }
         });
+
+
+        mKSIHRulesViewModel.getText().observe(this, s -> mFragmentKsihrulesBinding.textviewKsihRules.setText("Ksih rules"));
         return mFragmentKsihrulesBinding.getRoot();
     }
