+++ app/src/main/java/com/project/ksih_android/utility/Constants.java
@@ -9,4 +9,5 @@ public class Constants {
     public static String ON_BOARDING_KEY = "user_seen_onboarding_screen";
     public static String REMEMBER_ME = "user_checks_remember_me_checkbox";
     public static String SHARED_PREF_KEY = "KSIH-android-app";
+    public static String ALERT_DIALOG_PREF_KEY = "KSIH-android-app-alert_dialog-key";
 }
